backend/pyproject.toml
[tool.poetry]
name = "lg-webapp"
version = "0.1.0"
description = ""
authors = ["Woohyun Kim  <woodenkim92@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.7"
django-filter = "^23.2"
django-environ = "^0.11.2"
pillow = "^10.0.1"
coverage = "^7.3.1"
django-cors-headers = "^4.3.0"
djangorestframework-simplejwt = "^5.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
